Deduplicate json reader
