+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/tracee/UJsonTraceReader.scala
@@ -4,7 +4,7 @@ import at.forsyte.apalache.infra.passes.options.SourceOption
 import at.forsyte.apalache.infra.passes.options.SourceOption._
 import at.forsyte.apalache.io.itf.ItfToTla
 import at.forsyte.apalache.io.json.JsonDeserializationError
-import at.forsyte.apalache.io.json.impl.{UJsonRep, UJsonScalaFactory, UJsonToTlaViaBuilder}
+import at.forsyte.apalache.io.json.impl.{UJsonRep, UJsonScalaFactory, UJsonToTla}
 import at.forsyte.apalache.io.lir.TypeTagReader
 import at.forsyte.apalache.tla.imp.src.SourceStore
 import at.forsyte.apalache.tla.lir.oper.{TlaBoolOper, TlaOper}
@@ -26,7 +26,7 @@ import scala.util.{Failure, Success, Try}
  *   Jure Kukovec
  */
 class UJsonTraceReader(sourceStoreOpt: Option[SourceStore], tagReader: TypeTagReader) extends TraceReader[UJsonRep] {
-  private val builder = new UJsonToTlaViaBuilder(sourceStoreOpt)(tagReader)
+  private val builder = new UJsonToTla(sourceStoreOpt)(tagReader)
   private val itfToTla = new ItfToTla[UJsonRep](UJsonScalaFactory)
 
   type TraceUJson = TraceJson[UJsonRep]
