+++ tla-io/src/main/scala/at/forsyte/apalache/io/json/impl/UJsonToTla.scala
@@ -1,11 +1,12 @@
 package at.forsyte.apalache.io.json.impl
 
 import at.forsyte.apalache.io.json.JsonToTla
-import at.forsyte.apalache.tla.imp.src.SourceStore
 import at.forsyte.apalache.io.lir.TypeTagReader
+import at.forsyte.apalache.tla.imp.src.SourceStore
 
 /**
- * A json decoder, using the UJson representation
+ * @author
+ *   Jure Kukovec
  */
 class UJsonToTla(sourceStoreOpt: Option[SourceStore] = None)(implicit typeTagReader: TypeTagReader)
     extends JsonToTla[UJsonRep](UJsonScalaFactory, sourceStoreOpt)(typeTagReader)
