+++ tla-io/src/test/scala/at/forsyte/apalache/io/json/TestUJsonToTla.scala
@@ -4,14 +4,16 @@ import at.forsyte.apalache.io.json.impl.{DefaultTagReader, TlaToUJson, UJsonToTl
 import at.forsyte.apalache.io.lir.TlaType1PrinterPredefs
 import at.forsyte.apalache.tla.lir.UntypedPredefs._
 import at.forsyte.apalache.tla.lir._
-import at.forsyte.apalache.tla.lir.convenience.tla
-import at.forsyte.apalache.tla.lir.values.{TlaBoolSet, TlaIntSet, TlaNatSet, TlaStrSet}
+import at.forsyte.apalache.tla.lir.oper.TlaArithOper
+import at.forsyte.apalache.tla.types.tla
 import org.junit.runner.RunWith
-import org.scalacheck.Prop.{forAll, AnyOperators}
 import org.scalatest.funsuite.AnyFunSuite
 import org.scalatestplus.junit.JUnitRunner
 import org.scalatestplus.scalacheck.Checkers
 
+/**
+ * TODO: Implement generators for well-typed IR, then rewrite as PBT
+ */
 @RunWith(classOf[JUnitRunner])
 class TestUJsonToTla extends AnyFunSuite with Checkers {
   implicit val reader = DefaultTagReader
@@ -22,18 +24,15 @@ class TestUJsonToTla extends AnyFunSuite with Checkers {
 
   test("dec(enc(x)) =?= x") {
     val exs: Seq[TlaEx] = Seq(
-        tla.and(tla.name("x"), tla.bool(true)),
-        tla.ite(tla.name("p"), tla.name("A"), tla.name("B")),
+        tla.and(tla.name("x", BoolT1), tla.bool(true)),
+        tla.ite(tla.name("p", BoolT1), tla.name("A", ConstT1("X")), tla.name("B", ConstT1("X"))),
         tla.letIn(
-            tla.appOp(tla.name("A"), tla.int(1)),
-            tla
-              .declOp(
-                  "A",
-                  tla.plus(tla.name("p"), tla.int(1)),
-                  OperParam("p"),
-              )
-              .withTag(Untyped)
-              .asInstanceOf[TlaOperDecl],
+            tla.appOp(tla.name("A", OperT1(Seq(IntT1), IntT1)), tla.int(1)),
+            tla.decl(
+                "A",
+                tla.plus(tla.name("p", IntT1), tla.int(1)),
+                tla.param("p", IntT1),
+            ),
         ),
     )
 
@@ -43,8 +42,21 @@ class TestUJsonToTla extends AnyFunSuite with Checkers {
       assert(decEx == ex)
     }
 
+    // Typed builder should fail on these, and the exceptions should be rethrown as deserialization exceptions
+    val badExs: Seq[TlaEx] = Seq(
+        NameEx("a")(Untyped),
+        OperEx(TlaArithOper.plus, tla.int(1), tla.name("x", BoolT1))(Typed(IntT1)),
+    )
+
+    badExs.foreach { ex =>
+      val encEx = enc(ex)
+      assertThrows[JsonDeserializationError] {
+        dec.asTlaEx(encEx)
+      }
+    }
+
     val decls: Seq[TlaDecl] = Seq(
-        tla.declOp("X", tla.eql(tla.name("a"), tla.int(1)), OperParam("a")),
+        tla.decl("X", tla.eql(tla.name("a", IntT1), tla.int(1)), tla.param("a", IntT1)),
         TlaAssumeDecl(tla.eql(tla.int(1), tla.int(0))),
         TlaConstDecl("c"),
         TlaVarDecl("v"),
@@ -55,8 +67,8 @@ class TestUJsonToTla extends AnyFunSuite with Checkers {
     }
 
     val modules: Seq[TlaModule] = Seq(
-        new TlaModule("Empty", Seq.empty),
-        new TlaModule("Module", decls),
+        TlaModule("Empty", Seq.empty),
+        TlaModule("Module", decls),
     )
 
     modules.foreach { m =>
@@ -67,14 +79,12 @@ class TestUJsonToTla extends AnyFunSuite with Checkers {
   }
 
   test("Predef sets (see #1281)") {
-    val sets: Seq[ValEx] = Seq(
-        TlaIntSet,
-        TlaNatSet,
-        TlaBoolSet,
-        TlaStrSet,
-    ).map { v =>
-      ValEx(v).withTag(Untyped)
-    }
+    val sets: Seq[TlaEx] = Seq(
+        tla.intSet(),
+        tla.natSet(),
+        tla.booleanSet(),
+        tla.stringSet(),
+    )
 
     sets.foreach { s =>
       assert(dec.asTlaEx(enc(s)) == s)
@@ -82,52 +92,21 @@ class TestUJsonToTla extends AnyFunSuite with Checkers {
 
   }
 
-  test("TypeReader correctly reads valid type strings and fails on invalid type strings") {
-    val valid: Seq[String] = Seq(
-        "Untyped",
-        "Int",
-        "Set(Bool)",
-        "() => a -> b",
-        "(UT, Int) => [x: Str]",
-    )
-
-    val invalid: Seq[String] = Seq(
-        "",
-        "SomeType",
-        "Untyped()",
-        "-12",
-        "Set(true)",
-        "() > a -> b",
-        "(UT, Int) => [x: 9]",
-        "Typed[Str](\"cake\")",
-    )
-
-    // No throw
-    valid.foreach { s =>
-      DefaultTagReader.apply(s)
-    }
-    invalid.foreach { s =>
-      assertThrows[JsonDeserializationError] {
-        DefaultTagReader.apply(s)
-      }
-    }
-
-  }
-
-  test("Deserializing a serialized IR produces an equivalent IR") {
-    val gens: IrGenerators = new IrGenerators {
-      override val maxArgs: Int = 3
-    }
-    val operators =
-      gens.simpleOperators ++ gens.logicOperators ++ gens.arithOperators ++ gens.setOperators ++ gens.functionOperators ++ gens.actionOperators ++ gens.temporalOperators
-    val genDecl = gens.genTlaDeclButNotVar(gens.genTlaEx(operators)) _
-    val prop = forAll(gens.genTlaModuleWith(genDecl)) { module =>
-      val moduleJson = enc(module)
-      val moduleFromJson = dec.asTlaModule(moduleJson)
-
-      module =? moduleFromJson
-    }
-    check(prop, minSuccessful(500), sizeRange(7))
-  }
+  // TODO: Uncomment once generated IRs are well-typed
+//  test("Deserializing a serialized IR produces an equivalent IR") {
+//    val gens: IrGenerators = new IrGenerators {
+//      override val maxArgs: Int = 3
+//    }
+//    val operators =
+//      gens.simpleOperators ++ gens.logicOperators ++ gens.arithOperators ++ gens.setOperators ++ gens.functionOperators ++ gens.actionOperators ++ gens.temporalOperators
+//    val genDecl = gens.genTlaDeclButNotVar(gens.genTlaEx(operators)) _
+//    val prop = forAll(gens.genTlaModuleWith(genDecl)) { module =>
+//      val moduleJson = enc(module)
+//      val moduleFromJson = dec.asTlaModule(moduleJson)
+//
+//      module =? moduleFromJson
+//    }
+//    check(prop, minSuccessful(500), sizeRange(7))
+//  }
 
 }
